Company#getCompanyID() is now of type long, tests should take this into account
git-svn-id: https://activeobjects.dev.java.net/svn/activeobjects/trunk@555 69e892e8-b12f-0410-b7e1-b50061bc6545

+++ ActiveObjects/test/net/java/ao/db/hsqldb-create-table.sql
@@ -6,7 +6,7 @@ CREATE TABLE person (
     url VARCHAR UNIQUE NOT NULL,
     favoriteClass VARCHAR,
     height DOUBLE(32,6) DEFAULT 62.3,
-    companyID INTEGER,
+    companyID BIGINT,
     cool BOOLEAN DEFAULT TRUE,
     created TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT fk_person_companyid FOREIGN KEY (companyID) REFERENCES company(id),
