+++ ActiveObjects/test/net/java/ao/db/sqlserver-create-table.sql
@@ -6,7 +6,7 @@ CREATE TABLE person (
     url VARCHAR UNIQUE NOT NULL,
     favoriteClass VARCHAR,
     height DOUBLE(32,6) DEFAULT 62.3,
-    companyID INTEGER,
+    companyID BIGINT,
     cool INTEGER DEFAULT 1,
     created DATETIME DEFAULT GetDate(),
     CONSTRAINT fk_person_companyid FOREIGN KEY (companyID) REFERENCES company(id),
